server/requirements.txt
cryptography
Flask
python-dotenv